+++ duckscript_sdk/src/sdk/std/env/mod.rs
@@ -1,5 +1,6 @@
 mod cd;
 mod get;
+mod get_home_dir;
 mod pwd;
 mod set;
 mod unset;
@@ -15,6 +16,7 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
 
     commands.set(cd::create(&package))?;
     commands.set(get::create(&package))?;
+    commands.set(get_home_dir::create(&package))?;
     commands.set(pwd::create(&package))?;
     commands.set(set::create(&package))?;
     commands.set(unset::create(&package))?;